refactor(重构): 重构表格实现
去除引入的第三方表格生成依赖,重新实现,使得表格生成属性可细化至单元格。

+++ src/main/java/cn/meowy/pdf/utils/ColorUtils.java
@@ -32,7 +32,7 @@ public class ColorUtils {
             if (size == 1) {
                 return ColorUtil.getColor(color);
             } else if (size == 3) {
-                return ExUtils.execute(() -> new Color(NumberUtil.parseFloat(split.get(0)), NumberUtil.parseFloat(split.get(1)), NumberUtil.parseFloat(split.get(2))), "初始化颜色失败!");
+                return ExUtils.execute(() -> new Color(NumberUtil.parseInt(split.get(0)), NumberUtil.parseInt(split.get(1)), NumberUtil.parseInt(split.get(2))), "初始化颜色失败!");
             }
             throw new RuntimeException("色彩配置错误!");
         }
